feat(wasm): publish the lightweight compiler as the default entry
