+++ .github/workflows/deploy-docs.yml
@@ -70,7 +70,7 @@ jobs:
         # Built from rsvelte_lint_bindings, which links rsvelte_core's compiler +
         # svelte2tsx wasm exports transitively and adds the `lint` export that
         # powers the playground's Lint tab.
-        run: wasm-pack build crates/rsvelte_lint_bindings --out-dir ../../pkg --target web --no-default-features --features wasm
+        run: wasm-pack build crates/rsvelte_lint_bindings --out-dir ../../pkg-playground --target web --no-default-features --features wasm
 
       - name: Build WASM (formatter)
         # Powers the playground's `fmt` tool. Separate crate (rsvelte_fmt_wasm)
