+++ .gitignore
@@ -18,8 +18,8 @@ target
 /pkg
 # Formatter WASM build output (rsvelte_fmt_wasm → browser fmt tool)
 /pkg-fmt
-# Compiler-only WASM build output (rsvelte_compiler_wasm)
-/pkg-compiler
+# Playground WASM build output (compiler, lint and svelte2tsx)
+/pkg-playground
 
 # Language server: vendored lint wasm (built via build:wasm:lint-node) + bundle
 /apps/npm/language-server/vendor
