+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "crates/rsvelte_bench",
     "crates/rsvelte_check",
     "crates/rsvelte_compiler_wasm",
+    "crates/rsvelte_compiler_wasm_bindings",
     "crates/rsvelte_content_mapper",
     "crates/rsvelte_core",
     "crates/rsvelte_devtools",
