+++ package.json
@@ -4,13 +4,13 @@
 	"private": true,
 	"type": "module",
 	"scripts": {
-		"build:wasm": "pnpm run build:wasm:core && pnpm run build:wasm:fmt",
-		"build:wasm:core": "wasm-pack build crates/rsvelte_lint_bindings --out-dir ../../pkg --target web --release -- --features wasm --no-default-features",
+		"build:wasm": "pnpm run build:wasm:core && pnpm run build:wasm:playground && pnpm run build:wasm:fmt",
+		"build:wasm:core": "CARGO_PROFILE_RELEASE_OPT_LEVEL=z wasm-pack build crates/rsvelte_compiler_wasm --out-dir ../../pkg --target web --release",
+		"build:wasm:playground": "wasm-pack build crates/rsvelte_lint_bindings --out-dir ../../pkg-playground --target web --release -- --features wasm --no-default-features",
 		"build:wasm:lint-node": "wasm-pack build crates/rsvelte_lint_bindings --out-dir ../../apps/npm/language-server/vendor --target nodejs --release -- --features wasm --no-default-features",
 		"build:language-server": "pnpm run build:wasm:lint-node && pnpm --filter @rsvelte/language-server run build",
 		"build:vscode": "pnpm run build:language-server && pnpm --filter rsvelte run build",
 		"build:wasm:fmt": "wasm-pack build crates/rsvelte_fmt_wasm --out-dir ../../pkg-fmt --target web --release",
-		"build:wasm:compiler": "wasm-pack build crates/rsvelte_compiler_wasm --out-dir ../../pkg-compiler --target web --release",
 		"build:wasm:clean": "cargo clean && pnpm run build:wasm",
 		"clean:cache": "rm -rf apps/playground/node_modules/.vite",
 		"dev": "pnpm run build:wasm && pnpm run clean:cache && cd apps/playground && pnpm run dev",
@@ -82,8 +82,10 @@
 		"test:napi-options": "node scripts/dev/test-napi-compile-options.mjs",
 		"test:napi-csshash": "node scripts/dev/test-napi-css-hash.mjs",
 		"test:napi-preprocess-parity": "node scripts/dev/test-napi-preprocess-parity.mjs",
-		"test:wasm-compile": "node scripts/dev/test-wasm-compile-options.mjs",
+		"test:wasm-package": "node scripts/dev/test-wasm-package.mjs",
+		"test:wasm-browser": "node scripts/dev/test-wasm-browser.mjs",
 		"test:wasm-compiler-only": "node scripts/dev/test-wasm-compiler-only.mjs",
+		"test:wasm-compile": "node scripts/dev/test-wasm-compile-options.mjs",
 		"test:compare-cli": "node --test apps/npm/compiler/test/compare-cli.test.mjs",
 		"test:compare-cli:e2e": "node apps/npm/compiler/test/compare-cli-e2e.mjs",
 		"test:type-aware-lint": "node scripts/dev/test-type-aware-lint.mjs",
