+++ scripts/release/sync-version.mjs
@@ -1,38 +1,6 @@
 #!/usr/bin/env node
 // Sync each npm package version (managed by changesets) into the matching
 // Rust crate's `Cargo.toml` `[package].version` and the repo-root `Cargo.lock`.
-//
-// Why this exists:
-// - `@rsvelte/compiler` ← `crates/rsvelte_core` AND `crates/rsvelte_lint`:
-//   `@rsvelte/compiler` ships the wasm built from `crates/rsvelte_lint_bindings`
-//   (`build:wasm:core`), which re-exports the `rsvelte_core` compiler wasm API
-//   and the linter engine. The two runtime version strings baked into that wasm
-//   module resolve to CRATE versions: `rsvelte_core` backs the compiler
-//   `version()` export via its own `env!("CARGO_PKG_VERSION")`, and the
-//   bindings' `lint_version()` reads `rsvelte_lint::CRATE_VERSION` (this crate's
-//   version, not the bindings crate's). Keeping BOTH aligned with the release
-//   version keeps those strings honest. (The published `pkg/package.json`
-//   version itself is forced by `finalize-pkg.mjs`, which is what actually
-//   guards against a build-crate/version desync — but we still mirror both
-//   crates so the in-wasm version exports don't drift.)
-//   `crates/rsvelte_lint` must be mapped here too since `lint_version()` reports
-//   its version; without it `lint_version()` would report a stale `0.1.0`. The
-//   native `@rsvelte/lint` CLI (built from `crates/rsvelte_lint`, reporting
-//   `--version` from `CARGO_PKG_VERSION`) needs no separate mapping: it shares a
-//   `fixed` changeset group with `@rsvelte/compiler` (see
-//   `.changeset/config.json`), so it always bumps to the same version this rule
-//   already mirrors into the `rsvelte_lint` crate.
-// - `@rsvelte/fmt` ← `crates/rsvelte_fmt`: the `rsvelte-fmt` binary reports its
-//   version from `env!("CARGO_PKG_VERSION")` (clap `#[command(version)]`).
-//   Without this sync the crate stays at `0.1.0` no matter how many releases
-//   ship, so `rsvelte-fmt --version` would report a stale version that never
-//   matches the published `@rsvelte/fmt` package.
-// - `@rsvelte/language-server` ← `crates/rsvelte_language_server`: the native
-//   server reports `serverInfo.version` from `env!("CARGO_PKG_VERSION")`, which
-//   editors surface, so it must track the npm package it ships in.
-//
-// Each binary's `--version` must match the npm package it ships in.
-
 import { readFileSync, writeFileSync } from 'node:fs';
 import { fileURLToPath } from 'node:url';
 import { dirname, resolve } from 'node:path';
@@ -43,6 +11,16 @@ const repoRoot = resolve(here, '../..');
 // npm package.json (changeset-managed) → Rust crate to mirror into.
 // `lockName` is the crate's `name` in Cargo.lock.
 const MAPPINGS = [
+	{
+		npm: 'apps/npm/compiler/package.json',
+		cargoToml: 'crates/rsvelte_compiler_wasm_bindings/Cargo.toml',
+		lockName: 'rsvelte_compiler_wasm_bindings',
+	},
+	{
+		npm: 'apps/npm/compiler/package.json',
+		cargoToml: 'crates/rsvelte_compiler_wasm/Cargo.toml',
+		lockName: 'rsvelte_compiler_wasm',
+	},
 	{
 		npm: 'apps/npm/compiler/package.json',
 		cargoToml: 'crates/rsvelte/Cargo.toml',
@@ -59,7 +37,7 @@ const MAPPINGS = [
 		lockName: 'rsvelte_projection',
 	},
 	{
-		// The crate `build:wasm:core` actually builds into `pkg/` → `@rsvelte/compiler`.
+		// The playground export also contains the lint engine.
 		npm: 'apps/npm/compiler/package.json',
 		cargoToml: 'crates/rsvelte_lint/Cargo.toml',
 		lockName: 'rsvelte_lint',
